+++ src/pipecat/services/elevenlabs.py
@@ -388,7 +388,11 @@ class ElevenLabsTTSService(WordTTSService, WebsocketService):
     async def _keepalive_task_handler(self):
         while True:
             await asyncio.sleep(10)
-            await self._send_text("")
+            try:
+                await self._send_text("")
+            except websockets.ConnectionClosed as e:
+                logger.warning(f"{self} keepalive error: {e}")
+                break
 
     async def _send_text(self, text: str):
         if self._websocket: