+++ CHANGELOG.md
@@ -47,6 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an `RTVI` issue that was causing `bot-tts-text` messages to be sent
   before being processed by the output transport.
 
+- Fixed an issue[#1192] in 11labs where we are trying to reconnect/disconnect the
+  websocket connection even when the connection is already closed.
+
 ## [0.0.56] - 2025-02-06
 
 ### Changed